requirements.txt
Dlib-bin
numpy
opencv_contrib_python
opencv_python
pandas
Pillow
Pillow
plotly
plyer
streamlit
streamlit-aggrid
tensorflow
wget
pylibsrtp
streamlit-webrtc
